First exploratory programming for HTML conversion
We'll eventually take the Word doc, save it as a .mht file (HTML
embedded in a MIME wrapper) and use that to convert the data, because we
can actually figure out the indentation levels that way. As a bonus,
that will also get us a clean input file for the convert.py script,
where we don't have to play around with newlines, because the HTML
conversion from Word makes a proper distinction between wrapped lines
and the end of paragraphs.